+++ Plugin/CodeRebirth/src/Content/Enemies/Transporter.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using CodeRebirth.src.Util.Extensions;
+using CodeRebirthLib.Util.Pathfinding;
 using GameNetcodeStuff;
 using Unity.Netcode;
 using UnityEngine;
@@ -187,13 +188,13 @@ public class Transporter : CodeRebirthEnemyAI
         smartAgentNavigator.CheckPaths(candidateObjects, CheckIfNeedToChangeState);
     }
 
-    public void CheckIfNeedToChangeState(List<(GameObject gameObject, float distance)> args)
+    public void CheckIfNeedToChangeState(List<GenericPath<GameObject>> args)
     {
         int totalAmount = args.Count;
         if (totalAmount > 0)
         {
             Plugin.ExtendedLogging($"Transporter: Found {totalAmount} objects");
-            transportTarget = args[UnityEngine.Random.Range(0, totalAmount)].gameObject;
+            transportTarget = args[UnityEngine.Random.Range(0, totalAmount)].Generic;
             objectsToTransport.Remove(transportTarget);
             smartAgentNavigator.StopSearchRoutine();
             SwitchToBehaviourServerRpc((int)TransporterStates.Transporting);
@@ -246,7 +247,7 @@ public class Transporter : CodeRebirthEnemyAI
         }
     }
 
-    public void CheckIfCanReposition(List<(GameObject gameObject, float distance)> args)
+    public void CheckIfCanReposition(List<GenericPath<GameObject>> args)
     {
         if (transportTarget == null)
         {
@@ -259,7 +260,7 @@ public class Transporter : CodeRebirthEnemyAI
         if (args.Count > 0)
         {
             // Plugin.ExtendedLogging($"Transporter: Found {objects.Count} objects");
-            Vector3 currentEndDestination = args[UnityEngine.Random.Range(0, args.Count)].gameObject.transform.position;
+            Vector3 currentEndDestination = args[UnityEngine.Random.Range(0, args.Count)].Generic.transform.position;
             NavMesh.SamplePosition(currentEndDestination, out currentEndHit, 4f, NavMesh.AllAreas);
             if (repositioning)
             {